Fix failing sqlite VerifyMigrationTask with Hsql

+++ sqldelight-gradle-plugin/src/main/kotlin/app/cash/sqldelight/gradle/VerifyMigrationTask.kt
@@ -113,6 +113,7 @@ abstract class VerifyMigrationTask : SqlDelightWorkerTask() {
     }
 
     override fun execute() {
+      if (!environment.dialect.isSqlite) return
       parameters.workingDirectory.get().asFile.deleteRecursively()
 
       val catalog = createCurrentDb()
